Bismillah/requirements.txt
python-telegram-bot==21.6
supabase==2.9.1
openai==1.51.0
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
aiohttp==3.10.10
httpx==0.27.2
python-dotenv==1.0.1
pytz==2024.2
cryptography>=42.0.0
curl_cffi>=0.7.4
websockets>=12.0
